Rearrange and include plugin tests
