+++ backend/mirai-console/sandbox-tester/p1/src/p/Shadowed.kt
