--- backend/mirai-console/tester-plugins/AutoEnd/resources/META-INF/services/net.mamoe.mirai.console.plugin.jvm.KotlinPlugin
@@ -1,1 +1,0 @@
-autoend.Main
