--- backend/mirai-console/tester-plugins/p1/resources/META-INF/services/net.mamoe.mirai.console.plugin.jvm.KotlinPlugin
@@ -1,1 +1,0 @@
-p.s1.P
