--- backend/mirai-console/tester-plugins/p2/build.gradle.kts
@@ -1,21 +1,0 @@
-/*
- * Copyright 2019-2021 Mamoe Technologies and contributors.
- *
- * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
- * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
- *
- * https://github.com/mamoe/mirai/blob/dev/LICENSE
- */
-
-plugins {
-    kotlin("jvm")
-    kotlin("plugin.serialization")
-}
-
-kotlin {
-    explicitApiWarning()
-}
-
-dependencies {
-    implementation(project(":mirai-console-plugin-tester-common"))
-}
