--- backend/mirai-console/tester-plugins/p2/resources/META-INF/services/net.mamoe.mirai.console.plugin.jvm.KotlinPlugin
@@ -1,1 +1,0 @@
-p.s2.P
