--- backend/mirai-console/tester-plugins/p2/src/p/s2/P.kt
@@ -1,20 +1,0 @@
-/*
- * Copyright 2019-2021 Mamoe Technologies and contributors.
- *
- * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
- * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
- *
- * https://github.com/mamoe/mirai/blob/dev/LICENSE
- */
-
-package p.s2
-
-import net.mamoe.mirai.console.testing.BaseTestingPlugin
-import p.Shadowed
-import kotlin.test.assertEquals
-
-public object P : BaseTestingPlugin("p2") {
-    override fun onEnable0() {
-        assertEquals("Test S", Shadowed.v)
-    }
-}
