+++ tools/gradle-plugin/src/main/kotlin/VersionConstants.kt
@@ -10,6 +10,6 @@
 package net.mamoe.mirai.console.gradle
 
 internal object VersionConstants {
-    const val CONSOLE_VERSION = "2.7-M2-dev-1" // value is written here automatically during build
-    const val CORE_VERSION = "2.7-M2-dev-1" // value is written here automatically during build
+    const val CONSOLE_VERSION = "2.7-RC" // value is written here automatically during build
+    const val CORE_VERSION = "2.7-RC" // value is written here automatically during build
 }
